Add README for Android code (other than socket code). Add README instructions for pulling down and setting up the server. Ensure that Fragments aren't using actions to post transitions. Remove unnecessary action classes. Move floating_search_view query listener to ListPagingFragment where it exists in the direct view hierarchy. Make ListDiff implement List by delegation, making access cleaner, Rename moving package to navigating,
