+++ android/App/app/src/main/java/com/sdp15/goodb0i/view/BaseFragment.kt
@@ -12,4 +12,6 @@ abstract class BaseFragment : Fragment() {
 
     val baseActivity: MainActivity
         get() = activity as MainActivity
+
+
 }
