+++ android/App/app/src/main/java/com/sdp15/goodb0i/view/list/code/CodeViewModel.kt
@@ -2,18 +2,17 @@ package com.sdp15.goodb0i.view.list.code
 
 import androidx.annotation.StringRes
 import androidx.lifecycle.MutableLiveData
-import com.sdp15.goodb0i.view.BaseViewModel
 import com.sdp15.goodb0i.R
 import com.sdp15.goodb0i.data.store.Result
 import com.sdp15.goodb0i.data.store.lists.ListManager
-import com.sdp15.goodb0i.data.store.lists.ShoppingList
+import com.sdp15.goodb0i.view.BaseViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import org.koin.standalone.inject
 import timber.log.Timber
 
-class CodeViewModel : BaseViewModel<CodeViewModel.CodeAction>(),
+class CodeViewModel : BaseViewModel<Any>(),
     CodeFragment.PinInteractor {
 
     private val listManager: ListManager by inject()
@@ -31,12 +30,9 @@ class CodeViewModel : BaseViewModel<CodeViewModel.CodeAction>(),
                 val result = listManager.loadList(input.toLong())
                 if (result is Result.Success) {
                     Timber.i("Retrieved list ${result.data}")
-                    actions.postValue(
-                        CodeAction.ConfirmShoppingListAction(
-                            result.data
-                        )
-                    )
+                    transitions.postValue(CodeFragmentDirections.actionCodeFragmentToListConfirmationFragment(result.data))
                 } else {
+                    //TODO: Tell the uesr that the list doesn't exist, or can't be loaded
                     Timber.e("List load failure $result")
                 }
             }
@@ -45,8 +41,5 @@ class CodeViewModel : BaseViewModel<CodeViewModel.CodeAction>(),
         }
     }
 
-    sealed class CodeAction {
-        data class ConfirmShoppingListAction(val list: ShoppingList) : CodeAction()
-    }
 
 }
