--- android/README.MD
@@ -1,1 +1,0 @@
-## Android development tree
