+++ server/README.md
@@ -1,0 +1,10 @@
+# Setting up the server
+
+- Clone https://github.com/SDP15/goodb0i.git
+- Run git checkout -b [branch_name] origin/[branch_name]  (branch_name) is currently app_socket
+- Open /server/GoodB0iServer in IntelliJ
+- Click the "Import Gradle project" link in the bottom right
+- In the import window 
+  - Tick "use auto import"
+  - Tick "use default gradle wrapper"
+  - For the Gradle JVM option, select whichever version you have installed (Not the autofilled one)
